+++ core/src/main/java/com/github/games647/fastlogin/core/importer/Importer.java
@@ -1,12 +1,13 @@
 package com.github.games647.fastlogin.core.importer;
 
+import com.github.games647.fastlogin.core.AuthStorage;
+import java.sql.Connection;
 import java.sql.SQLException;
-
 import javax.sql.DataSource;
 
 public abstract class Importer {
 
-    public abstract int importData(DataSource source, DataSource target, String targetTable) throws SQLException;
+    public abstract int importData(Connection source, DataSource target, AuthStorage storage) throws SQLException;
 
     protected void closeQuietly(AutoCloseable closeable) {
         if (closeable != null) {